+++ src/main/java/net/geforcemods/securitycraft/screen/AlarmScreen.java
@@ -1,6 +1,7 @@
 package net.geforcemods.securitycraft.screen;
 
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -37,7 +38,6 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraftforge.client.gui.widget.ScrollPanel;
-import net.minecraftforge.registries.ForgeRegistries;
 
 public class AlarmScreen extends Screen {
 	private static final ResourceLocation GUI_TEXTURE = new ResourceLocation(SecurityCraft.MODID, "textures/gui/container/alarm.png");
@@ -145,9 +145,9 @@ public class AlarmScreen extends Screen {
 
 	public class SoundScrollList extends ScrollPanel {
 		private static final int SLOT_HEIGHT = 12, TEXT_OFFSET = 11;
-		public final List<SoundEvent> allSoundEvents = new ArrayList<>(ForgeRegistries.SOUND_EVENTS.getValues());
-		private final Map<SoundEvent, Component> soundEventKeys = new HashMap<>();
-		private List<SoundEvent> filteredSoundEvents;
+		public final List<ResourceLocation> allSoundEvents = minecraft.getSoundManager().getAvailableSounds().stream().sorted(Comparator.comparing(ResourceLocation::toLanguageKey)).toList();
+		private final Map<ResourceLocation, Component> soundEventKeys = new HashMap<>();
+		private List<ResourceLocation> filteredSoundEvents;
 		private SoundInstance playingSound;
 		private int selectedSoundIndex, contentHeight = 0;
 		private String previousSearchText = "";
@@ -234,7 +234,7 @@ public class AlarmScreen extends Screen {
 				else if (yStart > top + height)
 					break;
 
-				SoundEvent soundEvent = filteredSoundEvents.get(i);
+				ResourceLocation soundEvent = filteredSoundEvents.get(i);
 				Component name = getSoundEventComponent(soundEvent);
 
 				font.draw(pose, name, left + TEXT_OFFSET, yStart, 0xC6C6C6);
@@ -243,8 +243,8 @@ public class AlarmScreen extends Screen {
 			}
 		}
 
-		private Component getSoundEventComponent(SoundEvent soundEvent) {
-			return soundEventKeys.computeIfAbsent(soundEvent, t -> Utils.localize(soundEvent.getLocation().toLanguageKey()));
+		private Component getSoundEventComponent(ResourceLocation soundEvent) {
+			return soundEventKeys.computeIfAbsent(soundEvent, t -> Utils.localize(soundEvent.toLanguageKey()));
 		}
 
 		private void renderHighlightBox(int entryRight, Tesselator tesselator, int baseY, int slotBuffer, int slotIndex, int min) {
@@ -269,16 +269,16 @@ public class AlarmScreen extends Screen {
 
 		public void selectSound(int slotIndex) {
 			selectedSoundIndex = slotIndex;
-			AlarmScreen.this.selectSound(filteredSoundEvents.get(slotIndex).getLocation());
+			AlarmScreen.this.selectSound(filteredSoundEvents.get(slotIndex));
 		}
 
-		public void playSound(SoundEvent soundEvent) {
+		public void playSound(ResourceLocation soundEvent) {
 			SoundManager soundManager = Minecraft.getInstance().getSoundManager();
 
 			if (playingSound != null)
 				soundManager.stop(playingSound);
 
-			playingSound = SimpleSoundInstance.forUI(soundEvent, pitch, 1.0F);
+			playingSound = SimpleSoundInstance.forUI(SoundEvent.createFixedRangeEvent(soundEvent, 1.0F), pitch, 1.0F);
 			soundManager.play(playingSound);
 		}
 
@@ -286,7 +286,7 @@ public class AlarmScreen extends Screen {
 			//@formatter:off
 			filteredSoundEvents = new ArrayList<>(allSoundEvents
 					.stream()
-					.filter(e -> e.getLocation().toLanguageKey().contains(searchText))
+					.filter(e -> e.toLanguageKey().contains(searchText))
 					.toList());
 			//@formatter:on
 			recalculateContentHeight();
@@ -308,7 +308,7 @@ public class AlarmScreen extends Screen {
 		}
 
 		public void updateSelectedSoundIndex() {
-			selectedSoundIndex = Iterables.indexOf(filteredSoundEvents, se -> se.getLocation().equals(selectedSoundEvent));
+			selectedSoundIndex = Iterables.indexOf(filteredSoundEvents, se -> se.equals(selectedSoundEvent));
 		}
 
 		@Override