+++ Changelog.md
@@ -70,6 +70,7 @@
 - Fix: Reinforced Chiseled Deepslate cannot be crafted with Reinforced Cobbled Deepslate Slabs
 - Fix: The collision of Block Pocket Wall blocks gets desynchronized when activating a block pocket
 - Fix: The alarm screen closes when the inventory button is pressed in the search bar
+- Fix: The alarm can't play sounds added by resource packs
 - Removed: Ability to change the subpage in the SecurityCraft Manual using the arrow keys. This is still possible by scrolling while holding down CTRL
 - Removed: "Is always active" option of Keypads and Key Panels. The same behavior can be achieved by setting the "Signal length" option to 0 and activating the block
 - Misc.: Small optimizations to certain network and client performance